feat(serve-static): rename
