+++ packages/serve-static-middleware/src/ui5/middleware.ts
